feat: added initial qb64b code - broken
