+++ src/index.ts
@@ -2,3 +2,10 @@ export * from './lib/core.js'
 export { SAIDAlgoCodex } from './lib/digests.js'
 export { SAIDDex } from './lib/digests.js'
 export { Codex } from './lib/digests.js'
+export { versify } from './lib/versions.js'
+export { deversify } from './lib/versions.js'
+export { rematch } from './lib/versions.js'
+export { Rever } from './lib/versions.js'
+export { Vrsn_2_0 } from './lib/versions.js'
+export { Vrsn_1_0 } from './lib/versions.js'
+export { Version } from './lib/versions.js'
