+++ src/lib/base64.ts
@@ -1,3 +1,25 @@
+import { Buffer } from 'buffer';
+// base64url is supported by node Buffer, but not in buffer package for browser compatibility
+// https://github.com/feross/buffer/issues/309
+
+// Instead of using a node.js-only module and forcing us to polyfill the Buffer global,
+// we insert code from https://gitlab.com/seangenabe/safe-base64 here
+
+/**
+ * Encodes a buffer to a Base64URLSafe string
+ * @param buffer
+ */
+export function encodeBase64Url(buffer: Buffer) {
+  if (!Buffer.isBuffer(buffer)) {
+    throw new TypeError('`buffer` must be a buffer.');
+  }
+  return buffer
+    .toString('base64')
+    .replace(/\+/g, '-')
+    .replace(/\//g, '_')
+    .replace(/=+/, '');
+}
+
 /**
  * A map of all Base64URLSafe indexes to characters from 0-63. Includes A-Z, a-z, 0-9, -, and _
  */
@@ -115,3 +137,6 @@ export function intToB64(i: number, l = 1): string {
 
   return out
 }
+
+
+
